# Introduce `blockgen` as an optimization target