--- a/Cargo.toml
+++ b/Cargo.toml
@@ -47,6 +47,7 @@ lto = true        # Optimize our binary at link stage.
 codegen-units = 1 # Increases compile time but improves optimization alternatives.
 opt-level = 3     # Optimize with 'all' optimization flipped on. May produce larger binaries than 's' or 'z'.
 panic = "abort"
+debug = true # Enable debug symbols for profiling
 
 [profile.dev]
 panic = "abort"