--- a/examples/lading-otel-metrics.yaml
+++ b/examples/lading-otel-metrics.yaml
@@ -1,0 +1,31 @@
+generator:
+  - http:
+      seed: [2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47, 53,
+             59, 61, 67, 71, 73, 79, 83, 89, 97, 101, 103, 107, 109, 113, 127, 131]
+      headers:
+        content-type: "application/x-protobuf"
+      target_uri: "http://127.0.0.1:4318/v1/metrics"
+      bytes_per_second: "50 MiB"
+      parallel_connections: 5
+      method:
+        post:
+          maximum_prebuild_cache_size_bytes: "512 MiB"
+          variant:
+            opentelemetry_metrics:
+
+blackhole:
+  - http:
+      binding_addr: "127.0.0.1:9091"
+  - http:
+      binding_addr: "127.0.0.1:9092"
+  - http:
+      binding_addr: "127.0.0.1:4217"  # OTLP HTTP endpoint
+      body_variant: "nothing"
+  # - grpc:
+  #     binding_addr: "127.0.0.1:4318"  # OTLP gRPC endpoint
+
+target_metrics:
+  - prometheus:
+      uri: "http://127.0.0.1:8888/metrics"  # OTel collector metrics endpoint
+      tags:
+        component: "otel-collector"