### What does this PR do?

This commit introduces a new binary to the project 'blockgen' which is rigged to
    read a lading.yaml and generate the block cache for that config, doing nothing
    more. This allows us to use tools like hyperfine, perf etc to debug performance
    problems at a macro level, as opposed to a reliance directly on criterion
    benchmarks.

With the added example config we have from hyperfine:

```
Benchmark 1: ./target/release/blockgen examples/lading-otel-metrics.yaml
  Time (mean ± σ):      1.705 s ±  0.018 s    [User: 1.628 s, System: 0.070 s]
  Range (min … max):    1.689 s …  1.751 s    10 runs
```

### Motivation

REF SMPTNG-659