@scottopell 2025-05-08
isn't this pretty similar to https://github.com/DataDog/lading/blob/main/lading/src/bin/payloadtool.rs ?