@blt 2025-05-08
> isn't this pretty similar to https://github.com/DataDog/lading/blob/main/lading/src/bin/payloadtool.rs ?

Lol yep. I'd recorded "Generator failed to generate..." in my notes but didn't realize it was not a one-off tool that made it. I'll compare and adjust this PR or remove it.